+++ packages/koa-metrics/package.json
@@ -23,6 +23,7 @@
   },
   "homepage": "https://github.com/shopify/quilt/blob/master/packages/koa-metrics/README.md",
   "dependencies": {
+    "@shopify/statsd": "^1.0.0",
     "hot-shots": "^5.9.1",
     "tslib": "^1.9.3"
   },
